feat: 식단 및 음식 관련 엔티티 정의
Diet, DailyDiet, Food 엔티티 필드를 구현하였습니다.

+++ src/main/java/com/project/growfit/domain/Diet/entity/DailyDiet.java
@@ -8,7 +8,6 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +36,7 @@ public class DailyDiet extends BaseEntity {
 
     @Column(nullable = false)
     @JoinColumn(name = "total_calorie")
-    private double totalCalorie;
+    private long totalCalorie;
 
     @OneToMany(mappedBy = "dailyDiet", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Diet> diets = new ArrayList<>();
